+++ src/systems.c
@@ -206,7 +206,7 @@ void SPlayerInputUpdate(Scene* scene, usize entity)
             player->jumping = true;
             kinetic->velocity.y = -player->jumpVelocity;
 
-            int particleCount = GetRandomValue(10, 25);
+            u16 particleCount = GetRandomValue(10, 25);
             Event particleEvent;
             EventCloudParticleInit(&particleEvent, entity, particleCount);
             SceneRaiseEvent(scene, &particleEvent);
